+++ dao-impl/src/main/java/jm/stockx/api/dao/ItemInfoDaoImpl.java
@@ -298,12 +298,12 @@ public class ItemInfoDaoImpl extends AbstractDAO<ItemInfo, Long> implements Item
     }
 
     @Override
-    public Long getCountOfUserBuyingByUserId(Long userId) {
+    public int getCountOfUserBuyingByUserId(Long userId) {
         return entityManager.createQuery("" +
                 "SELECT COUNT(i.buyingInfo.id) " +
                 "FROM ItemInfo i " +
                 "WHERE i.buyingInfo.id = :id " +
-                "AND NOT i.buyingInfo.status = 'CANCELED'", Long.class)
+                "AND NOT i.buyingInfo.status = 'CANCELED'", Integer.class)
                 .setParameter("id", userId)
                 .getSingleResult();
     }
